requirements.txt
python-dotenv
streamlit
guardrails-ai
